+++ prisma/migrations/20251207181932_drop_ai_token_usage_total_tokens/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `totalTokens` on the `AiTokenUsage` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE "AiTokenUsage" DROP COLUMN "totalTokens";